Finish separating build script from update logic
