+++ crates/update-bindings/Cargo.toml
@@ -6,8 +6,6 @@ publish = false
 
 [dependencies]
 regex = "1.5.4"
-serde = { version = "1.0", features = [ "derive" ] }
-serde_json = "1.0"
 thiserror = "1.0.26"
 
 [dependencies.windows-bindgen]
