+++ crates/update-bindings/build.rs
@@ -1,0 +1,2 @@
+// This build script is only here so the OUT_DIR environment variable will be set at compile time.
+fn main() {}
